+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/DefaultEntityApplicationPanelBuilder.java
@@ -333,8 +333,8 @@ final class DefaultEntityApplicationPanelBuilder<M extends SwingEntityApplicatio
 	}
 
 	private void enableLookAndFeel() {
-		String className = validateLookAndFeel(lookAndFeelClassName());
 		try {
+			String className = lookAndFeelClassName();
 			findLookAndFeelProvider(className)
 							.orElse(lookAndFeelProvider(new LookAndFeelInfo("LookAndFeel", className)))
 							.enable();
@@ -538,18 +538,6 @@ final class DefaultEntityApplicationPanelBuilder<M extends SwingEntityApplicatio
 		if (!(exception instanceof CancelException)) {
 			Window focusOwnerParentWindow = parentWindow(getCurrentKeyboardFocusManager().getFocusOwner());
 			displayExceptionDialog(exception, focusOwnerParentWindow == null ? applicationFrame : focusOwnerParentWindow);
-		}
-	}
-
-	private static String validateLookAndFeel(String className) {
-		try {
-			Class.forName(className);
-
-			return className;
-		}
-		catch (ClassNotFoundException e) {
-			LOG.error("Exception while enabling Look and Feel", e);
-			throw new RuntimeException("LookAndFeel class not found: " + className, e);
 		}
 	}